implemented toggle modal with useContext
