+++ frontend/src/components/DesktopPage/index.js
@@ -3,6 +3,8 @@ import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Switch, Route, Redirect } from "react-router-dom";
 
+import TagModalProvider from "../../context/TagModalContext";
+
 import * as notebooksActions from "../../store/notebooks";
 import * as notesActions from "../../store/notes";
 import * as tagsActions from "../../store/tags";
@@ -33,24 +35,26 @@ const DesktopPage = () => {
 
 	return (
 		<div className="desktop-container">
-			<Navigation sessionUser={sessionUser} />
-			<Switch>
-				<Route path="/desktop">
-					<Desktop />
-				</Route>
-				<Route path="/notes">
-					<NotesPage />
-				</Route>
-				<Route path="/notebooks">
-					<NotebooksPage />
-				</Route>
-				<Route path={["/tags", "/tags/:tagId"]}>
-					<TagsPage />
-				</Route>
-				<Route path="/trash">
-					<TrashPage />
-				</Route>
-			</Switch>
+			<TagModalProvider>
+				<Navigation sessionUser={sessionUser} />
+				<Switch>
+					<Route path="/desktop">
+						<Desktop />
+					</Route>
+					<Route path="/notes">
+						<NotesPage />
+					</Route>
+					<Route path="/notebooks">
+						<NotebooksPage />
+					</Route>
+					<Route path={["/tags", "/tags/:tagId"]}>
+						<TagsPage />
+					</Route>
+					<Route path="/trash">
+						<TrashPage />
+					</Route>
+				</Switch>
+			</TagModalProvider>
 		</div>
 	);
 };
