+++ frontend/src/components/Navigation/index.js
@@ -1,3 +1,5 @@
+import { useRef } from "react";
+
 import Home from "./home.js";
 import User from "./user.js";
 import Notes from "./notes.js";
