+++ frontend/src/css/modals.css
@@ -4,7 +4,7 @@
 	position: fixed;
 	width: 100vw;
 	height: 100vh;
-	z-index: 1;
+	z-index: 2;
 	display: flex;
 	justify-content: center;
 	align-items: center;
