+++ frontend/src/css/tags.css
@@ -30,3 +30,7 @@
 	font-size: 12px;
 	padding: 2px 7px;
 }
+
+.tag i {
+	filter: brightness(200%);
+}
